fix: fix empty/wrong DML parameters required for AWLAN_Node
In order to fix the issue with OpenSync connectivity to the cloud, we need to adjust DML values.
Currently returned values are:
   Device.DeviceInfo.ModelName  - Cougar Run
   Device.DeviceInfo.X_RDKCENTRAL-COM_CMTS_MAC - 00:00:00:00:00:00

Instead, the values should be:
   Device.DeviceInfo.ModelName - PUMA7_CGR
   Device.DeviceInfo.X_RDKCENTRAL-COM_CMTS_MAC - should return CM WAN MAC address

These values should be adjusted for OpenSync flavors,
but if it is a generic issue, it can be extended to non-OpenSync flavors also.

Signed-off-by: Vyacheslav Salnikov <v.salnikov@inango-systems.com>